chore(naming): rename src/mocks/ -> src/extensions/
It's going to contain more than just mocks pretty soon!
